GA-138: PRM : Generate Role Assignments and Process 6 - Refresh User - Single User Mode and new PRM endpoint
