+++ src/integrationTest/java/uk/gov/hmcts/reform/orgrolemapping/controller/RefreshControllerIntegrationTest.java
@@ -1,8 +1,6 @@
 package uk.gov.hmcts.reform.orgrolemapping.controller;
 
-import feign.FeignException;
-import feign.Request;
-import feign.RequestTemplate;
+
 import org.jetbrains.annotations.NotNull;
 
 import org.junit.jupiter.api.BeforeEach;
@@ -32,20 +30,16 @@ import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.context.WebApplicationContext;
 import uk.gov.hmcts.reform.orgrolemapping.apihelper.Constants;
-import uk.gov.hmcts.reform.orgrolemapping.controller.advice.exception.ResourceNotFoundException;
-import uk.gov.hmcts.reform.orgrolemapping.controller.advice.exception.ServiceException;
 import uk.gov.hmcts.reform.orgrolemapping.controller.advice.exception.UnauthorizedServiceException;
 import uk.gov.hmcts.reform.orgrolemapping.controller.utils.MockUtils;
 import uk.gov.hmcts.reform.orgrolemapping.controller.utils.WiremockFixtures;
 import uk.gov.hmcts.reform.orgrolemapping.data.RefreshJobEntity;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.Appointment;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.CaseWorkerProfilesResponse;
-import uk.gov.hmcts.reform.orgrolemapping.domain.model.GetRefreshUsersResponse;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.JudicialBooking;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.JudicialBookingResponse;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.JudicialProfile;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.JudicialRefreshRequest;
-import uk.gov.hmcts.reform.orgrolemapping.domain.model.RefreshUser;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.RoleAssignmentRequestResource;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.UserRequest;
 import uk.gov.hmcts.reform.orgrolemapping.domain.service.RequestMappingService;
@@ -56,7 +50,6 @@ import uk.gov.hmcts.reform.orgrolemapping.feignclients.PRDFeignClient;
 import uk.gov.hmcts.reform.orgrolemapping.feignclients.RASFeignClient;
 import uk.gov.hmcts.reform.orgrolemapping.helper.AssignmentRequestBuilder;
 import uk.gov.hmcts.reform.orgrolemapping.helper.IntTestDataBuilder;
-import uk.gov.hmcts.reform.orgrolemapping.helper.TestDataBuilder;
 import uk.gov.hmcts.reform.orgrolemapping.launchdarkly.FeatureConditionEvaluator;
 import uk.gov.hmcts.reform.orgrolemapping.util.SecurityUtils;
 
@@ -67,7 +60,6 @@ import java.sql.ResultSet;
 import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -93,9 +85,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static uk.gov.hmcts.reform.orgrolemapping.apihelper.Constants.ABORTED;
 import static uk.gov.hmcts.reform.orgrolemapping.apihelper.Constants.COMPLETED;
 import static uk.gov.hmcts.reform.orgrolemapping.apihelper.Constants.FAILED_ROLE_REFRESH;
-import static uk.gov.hmcts.reform.orgrolemapping.domain.service.ProfessionalRefreshOrchestrator.EXPECTED_SINGLE_PRD_USER;
-import static uk.gov.hmcts.reform.orgrolemapping.domain.service.ProfessionalRefreshOrchestrator.NO_ACCESS_TYPES_FOUND;
-import static uk.gov.hmcts.reform.orgrolemapping.domain.service.ProfessionalRefreshOrchestrator.PRD_USER_NOT_FOUND;
 import static uk.gov.hmcts.reform.orgrolemapping.v1.V1.Error.UNAUTHORIZED_SERVICE;
 
 @TestPropertySource(properties = {
@@ -613,79 +602,6 @@ public class RefreshControllerIntegrationTest extends BaseTestIntegration {
                         .value(containsString("The input parameter: \"abc-123$\", "
                                 + "does not comply with the required pattern")))
                 .andReturn();
-    }
-
-    @Test
-    @Sql(executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD,
-        scripts = {"classpath:sql/insert_user_refresh_queue_138.sql"})
-    public void shouldProcessProfessionalRefreshRequest() throws Exception {
-        doReturn(ResponseEntity.status(HttpStatus.OK).body(TestDataBuilder.buildRefreshUsersResponse("1234")))
-            .when(prdFeignClient).getRefreshUsers(any());
-
-        mockMvc.perform(post(PROFESSIONAL_REFRESH_URL + "?userId=1234")
-                .contentType(JSON_CONTENT_TYPE)
-                .headers(getHttpHeaders()))
-            .andExpect(status().isOk())
-            .andExpect(jsonPath("$.Message").value(containsString(Constants.SUCCESS_ROLE_REFRESH)))
-            .andReturn();
-    }
-
-    @Test
-    public void shouldRejectProfessionalRefreshRequest_withoutUserId() throws Exception {
-        mockMvc.perform(post(PROFESSIONAL_REFRESH_URL)
-                .contentType(JSON_CONTENT_TYPE)
-                .headers(getHttpHeaders()))
-            .andExpect(status().isBadRequest())
-            .andReturn();
-    }
-
-    @Test
-    public void shouldErrorProfessionalRefreshRequest_whenNoAccessTypesInDB() throws Exception {
-        doReturn(ResponseEntity.status(HttpStatus.OK).body(TestDataBuilder.buildRefreshUsersResponse("1234")))
-            .when(prdFeignClient).getRefreshUsers(any());
-
-        MvcResult result = mockMvc.perform(post(PROFESSIONAL_REFRESH_URL + "?userId=1234")
-                .contentType(JSON_CONTENT_TYPE)
-                .headers(getHttpHeaders()))
-            .andExpect(status().isInternalServerError())
-            .andReturn();
-
-        assertTrue(result.getResolvedException() instanceof ServiceException);
-        assertEquals(NO_ACCESS_TYPES_FOUND, result.getResolvedException().getMessage());
-    }
-
-    @Test
-    public void shouldErrorProfessionalRefreshRequest_whenNoPRDUserFound() throws Exception {
-        Request request = Request.create(Request.HttpMethod.GET, "url", new HashMap<>(), null, new RequestTemplate());
-        doThrow(new FeignException.NotFound("Not Found", request, null, null))
-            .when(prdFeignClient).getRefreshUsers(any());
-
-        MvcResult result = mockMvc.perform(post(PROFESSIONAL_REFRESH_URL + "?userId=1234")
-                .contentType(JSON_CONTENT_TYPE)
-                .headers(getHttpHeaders()))
-            .andExpect(status().isNotFound())
-            .andReturn();
-
-        assertTrue(result.getResolvedException() instanceof ResourceNotFoundException);
-        assertEquals(String.format(Constants.RESOURCE_NOT_FOUND + " " + PRD_USER_NOT_FOUND, "1234"),
-            result.getResolvedException().getMessage());
-    }
-
-    @Test
-    public void shouldErrorProfessionalRefreshRequest_whenMultipleUsersReturnedFromPRD() throws Exception {
-        GetRefreshUsersResponse getRefreshUsersResponse = TestDataBuilder.buildRefreshUsersResponse("1234");
-        getRefreshUsersResponse.getUsers().add(new RefreshUser());
-        doReturn(ResponseEntity.status(HttpStatus.OK).body(getRefreshUsersResponse))
-            .when(prdFeignClient).getRefreshUsers(any());
-
-        MvcResult result = mockMvc.perform(post(PROFESSIONAL_REFRESH_URL + "?userId=1234")
-                .contentType(JSON_CONTENT_TYPE)
-                .headers(getHttpHeaders()))
-            .andExpect(status().isInternalServerError())
-            .andReturn();
-
-        assertTrue(result.getResolvedException() instanceof ServiceException);
-        assertEquals(String.format(EXPECTED_SINGLE_PRD_USER, "1234", "2"), result.getResolvedException().getMessage());
     }
 
     private ResponseEntity<JudicialBookingResponse> buildJudicialBookingsResponse(String... userIds) {
