+++ src/integrationTest/resources/sql/delete_user_refresh_queue.sql
@@ -1,0 +1,5 @@
+DELETE FROM access_types;
+DELETE FROM user_refresh_queue;
+
+
+
